Projekti/Projekat/services/analytics/requirements.txt
click==8.1.3
colorama==0.4.4
Flask==2.1.2
grpcio==1.46.3
grpcio-tools==1.46.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
paho-mqtt==1.6.1
protobuf==3.20.1
pymongo==4.1.1
six==1.16.0
Werkzeug==2.1.2
